% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_yes.R
\name{any_yes}
\alias{any_yes}
\title{Title}
\usage{
any_yes(.data, variable, varcollapse)
}
\arguments{
\item{.data}{Dataframe that you want to create the variable in}

\item{variable}{The variable name that you want to give to collapsed variable}

\item{varcollapse}{The list of variables that you want to use to create the variable mentioned}
}
\value{
Dataframe with the newly created variable
}
\description{
Title
}
\examples{
df <- data.frame (id = 1:10,
var1 = factor(round(runif(20, 1, 3)),labels = c("Yes", "No", "N/A")),
var2 = factor(round(runif(20, 1, 3)),labels = c("Yes", "No", "N/A")),
var3 = factor(round(runif(20, 1, 3)),labels = c("Yes", "No", "N/A")))
df |> any_yes("ouctome", c("var1", "var2", "var3"))
}
