% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anymatch.R
\name{anymatch}
\alias{anymatch}
\title{Create variable from set of variables whose values are not missing}
\usage{
anymatch(vars)
}
\arguments{
\item{vars}{A set of variables that you want to create your variable}
}
\value{
A binary character vector with Yes or No
}
\description{
Create variable from set of variables whose values are not missing
}
\examples{
df <- data.frame (id = 1:10,
var1 = sample(c(1:2,NA), 10, replace=TRUE),
var2 = sample(c(1:2,NA), 10, replace=TRUE),
var3 = sample(c(3:5, NA), 10, replace=TRUE))
df |> dplyr::mutate(v4 = anymatch("var"))
}
