% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel}
\alias{export_excel}
\title{Export dataframe to excel}
\usage{
export_excel(data, fname)
}
\arguments{
\item{data}{Dataframe you want to export}

\item{fname}{File name of your exported dataframe}
}
\value{
A nicely formated excel file
}
\description{
Export dataframe to excel
}
\examples{
export_excel(mtcars, "mtcars")
}
