% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_keep.R
\name{labels_keep}
\alias{labels_keep}
\title{Keep the variable labels}
\usage{
labels_keep(data)
}
\arguments{
\item{data}{Dataframe that you want to preserve its variable labels}
}
\value{
A list of variable labels from the dataframe
}
\description{
Keep the variable labels
}
\examples{
df <- data.frame(a1=floor(rnorm(10, 15, 5)),
a2 = gl(2, 2, 10))
attr(df[['a1']], 'label') <- 'Age'
attr(df[['a2']], 'label') <- 'Gender'
labels_keep(df)
}
